+++ Backend/VFatFinger.java
@@ -2,9 +2,9 @@
 import java.lang.Math;
 
 public class VFatFinger implements ICheck {
-    int vma;
-    int lastVma;
-    float a = 0.18;   //effect of past average on vma
+    double vma;
+    double lastVma;
+    double a = 0.18;   //effect of past average on vma
     int n = 2;      //sensitivity of error detection
     int channel;    //historical/live data - whichever the check is for
 
@@ -27,7 +27,7 @@ public class VFatFinger implements ICheck {
         if( (stock.getSize() > lastVma*(Math.pow(10, n))) || ( (stock.getSize() < lastVma+(pow(10, 0-n))) ) ) {
             //then there has been a price ff error
             //calculate severity
-            int severity = (stock.getSize() * 100) / lastVma;
+            double severity = (stock.getSize() * 100) / lastVma;
             //send anomaly
             FFAnomaly anomaly = new FFAnomaly(client.getCounter(), channel, stock, severity, lastVma, 0);
             return anomaly;
